rename typings to ts
