+++ packages/cli-kit/src/public/common/ts/deep-required.ts
